+++ src/gen_why3.ml
@@ -1832,6 +1832,7 @@ let rec map_mterm m ctx (mt : M.mterm) : loc_term =
       let e = map_mterm m ctx d in
       let l = loc_term (Tnil gListAs) in
       Tapp( loc_term (Tvar "_mk_operation"),[a; e; l])
+    | Mmakeevent _ -> error_not_translated "Mmakeevent"
     | Mcreatecontract _   -> error_not_translated "Mcreatecontract"
 
     (* literals *)