+++ src/model.ml
@@ -253,6 +253,7 @@ type ('id, 'term) mterm_node  =
   (* operation *)
   | Moperations
   | Mmakeoperation    of 'term * 'term * 'term  (* value * address * args *)
+  | Mmakeevent        of type_ * 'id * 'term    (* type * id * arg *)
   | Mcreatecontract   of michelson_struct * 'term * 'term * 'term  (* value * option key_hash * address * init storage *)
   (* literals *)
   | Mint              of Core.big_int
@@ -1327,7 +1328,8 @@ let operations = mk_mterm Moperations (tlist toperation)
 
 let mk_get_entrypoint (ty: type_) (entry_name: lident) (addr : mterm) : mterm =
   mk_mterm (Mgetentrypoint (ty, entry_name, addr)) (toption (tcontract ty))
-let mk_mkoperation a b c = mk_mterm (Mmakeoperation (a, b, c)) tunit
+let mk_mkoperation a b c = mk_mterm (Mmakeoperation (a, b, c)) toperation
+let mk_mkevent     a b c = mk_mterm (Mmakeevent (a, b, c)) toperation
 let mk_transfer_op op = mk_mterm (Mtransfer (TKoperation op)) tunit
 
 (* -------------------------------------------------------------------- *)
@@ -1583,6 +1585,7 @@ let cmp_mterm_node
     (* operation *)
     | Moperations, Moperations                                                         -> true
     | Mmakeoperation (v1, d1, a1), Mmakeoperation (v2, d2, a2)                         -> cmp v1 v2 && cmp d1 d2 && cmp a1 a2
+    | Mmakeevent (t1, id1, a1), Mmakeevent (t2, id2, a2)                               -> cmp_type t1 t2 && cmp_lident id1 id2 && cmp a1 a2
     | Mcreatecontract (ms1, d1, a1, si1), Mcreatecontract (ms2, d2, a2, si2)           -> cmp_michelson_struct ms1 ms2 && cmp d1 d2 && cmp a1 a2 && cmp si1 si2
     (* literals *)
     | Mint v1, Mint v2                                                                 -> Big_int.eq_big_int v1 v2
@@ -2047,6 +2050,7 @@ let map_term_node_internal (fi : ident -> ident) (g : 'id -> 'id) (ft : type_ ->
   (* operation *)
   | Moperations                    -> Moperations
   | Mmakeoperation (v, d, a)       -> Mmakeoperation (f v, f d, f a)
+  | Mmakeevent (t, id, a)          -> Mmakeevent (ft t, g id, f a)
   | Mcreatecontract (ms, d, a, si) -> Mcreatecontract (ms, f d, f a, f si)
   (* literals *)
   | Mint v                         -> Mint v
@@ -2508,6 +2512,7 @@ let fold_term (f : 'a -> ('id mterm_gen) -> 'a) (accu : 'a) (term : 'id mterm_ge
   (* operation *)
   | Moperations                           -> accu
   | Mmakeoperation (v, d, a)              -> f (f (f accu v) d) a
+  | Mmakeevent (_, _, a)                  -> f accu a
   | Mcreatecontract (_, d, a, si)         -> f (f (f accu d) a) si
   (* literals *)
   | Mint _                                -> accu
@@ -3022,6 +3027,10 @@ let fold_map_term
     let de, da = f va d in
     let ae, aa = f da a in
     g (Mmakeoperation (ve, de, ae)), aa
+
+  | Mmakeevent (t, id, a) ->
+    let ae, aa = f accu a in
+    g (Mmakeevent (t, id, ae)), aa
 
   | Mcreatecontract (ms, d, a, si) ->
     let de, da = f accu d in