+++ CHANGES.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
   - Offchain views support (`-t offchain-views`)
+  - `make_event` builtin
 ### Fixed
   - Generate error message for side effect instructions in functions and views